Add team request model and routes
Successfully add team invite Requests as a sub document of the users
model. Currently a user can pull up their specific information and it
will shows an array with all pending requests. The idea is that when a
user sends a team invite, the user themselves gets a item added to
their requests array having the desired member as the Requester,
themselves as the recipient, and a status of Requested, whereas the
Recipient will have that same item but with the team creator as the
Requester, themselves as the Recipient, and a status of Pending.